fix: change test to retry after a successful write
Before we were checking the time right after startHook. The issue with
that is that there is variability in how much time it takes to call
updateHook after startHook (~200ms). We avoided that by writing once
before testing for the timeout, as time between updateHooks calls are
exactly 3s.

Another source of time variability is how long the first
modbus_expect_execution takes to finish due to the time it takes dealing
with the many modbus replies/requests. Because of that, tic is behind
the actual timeout timer start by ~0.250ms (without cpu restrictions).

We account for that variability by adding a threshold of 0.5s in the
time it takes to emit the timeout event.